# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: LFX V2 - Query Service
    description: Query indexed resources
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /query/orgs:
        get:
            tags:
                - query-svc
            summary: query-orgs query-svc
            description: Locate a single organization by name or domain.
            operationId: query-svc#query-orgs
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: true
                  type: string
                  enum:
                    - "1"
                - name: name
                  in: query
                  description: Organization name
                  required: false
                  type: string
                  minLength: 1
                - name: domain
                  in: query
                  description: Organization domain or website URL
                  required: false
                  type: string
                  pattern: ^[a-zA-Z0-9][a-zA-Z0-9-_.]*[a-zA-Z0-9]*\.[a-zA-Z]{2,}$
                - name: Authorization
                  in: header
                  description: Token
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Organization'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/NotFoundError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /query/orgs/suggest:
        get:
            tags:
                - query-svc
            summary: suggest-orgs query-svc
            description: Get organization suggestions for typeahead search based on a query.
            operationId: query-svc#suggest-orgs
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: true
                  type: string
                  enum:
                    - "1"
                - name: query
                  in: query
                  description: Search query for organization suggestions
                  required: true
                  type: string
                  minLength: 1
                - name: Authorization
                  in: header
                  description: Token
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/QuerySvcSuggestOrgsResponseBody'
                        required:
                            - suggestions
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /query/resources:
        get:
            tags:
                - query-svc
            summary: query-resources query-svc
            description: Locate resources by their type or parent, or use typeahead search to query resources by a display name or similar alias.
            operationId: query-svc#query-resources
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: true
                  type: string
                  enum:
                    - "1"
                - name: name
                  in: query
                  description: Resource name or alias; supports typeahead
                  required: false
                  type: string
                  minLength: 1
                - name: parent
                  in: query
                  description: Parent (for navigation; varies by object type)
                  required: false
                  type: string
                  pattern: ^[a-zA-Z]+:[a-zA-Z0-9_-]+$
                - name: type
                  in: query
                  description: Resource type to search
                  required: false
                  type: string
                - name: tags
                  in: query
                  description: Tags to search (varies by object type)
                  required: false
                  type: array
                  items:
                    type: string
                  collectionFormat: multi
                - name: sort
                  in: query
                  description: Sort order for results
                  required: false
                  type: string
                  default: name_asc
                  enum:
                    - name_asc
                    - name_desc
                    - updated_asc
                    - updated_desc
                - name: page_token
                  in: query
                  description: Opaque token for pagination
                  required: false
                  type: string
                - name: Authorization
                  in: header
                  description: Token
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/QuerySvcQueryResourcesResponseBody'
                        required:
                            - resources
                    headers:
                        Cache-Control:
                            description: Cache control header
                            type: string
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/BadRequestError'
                        required:
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/InternalServerError'
                        required:
                            - message
                "503":
                    description: Service Unavailable response.
                    schema:
                        $ref: '#/definitions/ServiceUnavailableError'
                        required:
                            - message
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /query/resources/count:
        get:
            tags:
                - query-svc
            summary: query-resources-count query-svc
            description: Count matching resources by query.
            operationId: query-svc#query-resources-count
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  required: true
                  type: string
                  enum:
                    - "1"
                - name: name
                  in: query
                  description: Resource name or alias; supports typeahead
                  required: false
                  type: string
                  minLength: 1
                - name: parent
                  in: query
                  description: Parent (for navigation; varies by object type)
                  required: false
                  type: string
                - name: type
                  in: query
                  description: Resource type to search
                  required: false
                  type: string
                - name: tags
                  in: query
                  description: Tags to search (varies by object type)
                  required: false
                  type: array
                  items:
                    type: string
                  collectionFormat: multi
                - name: Authorization
                  in: header
                  description: JWT token issued by Heimdall
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/QuerySvcQueryResourcesCountResponseBody'
                        required:
                            - count
                            - has_more
                    headers:
                        Cache-Control:
                            description: Cache control header
                            type: string
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/QuerySvcQueryResourcesCountBadRequestResponseBody'
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
definitions:
    BadRequestError:
        title: BadRequestError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The request was invalid.
        description: Bad request
        example:
            message: The request was invalid.
        required:
            - message
    InternalServerError:
        title: InternalServerError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: An internal server error occurred.
        description: Internal server error
        example:
            message: An internal server error occurred.
        required:
            - message
    NotFoundError:
        title: NotFoundError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The requested resource was not found.
        description: Not found
        example:
            message: The requested resource was not found.
        required:
            - message
    Organization:
        title: Organization
        type: object
        properties:
            domain:
                type: string
                description: Organization domain
                example: linuxfoundation.org
            employees:
                type: string
                description: Employee count or range
                example: 100-499
            industry:
                type: string
                description: Organization industry classification
                example: Non-Profit
            name:
                type: string
                description: Organization name
                example: Linux Foundation
            sector:
                type: string
                description: Business sector classification
                example: Technology
        example:
            domain: linuxfoundation.org
            employees: 100-499
            industry: Non-Profit
            name: Linux Foundation
            sector: Technology
    OrganizationSuggestion:
        title: OrganizationSuggestion
        type: object
        properties:
            domain:
                type: string
                description: Organization domain
                example: linuxfoundation.org
            logo:
                type: string
                description: Organization logo URL
                example: https://example.com/logo.png
            name:
                type: string
                description: Organization name
                example: Linux Foundation
        description: An organization suggestion for the search.
        example:
            domain: linuxfoundation.org
            logo: https://example.com/logo.png
            name: Linux Foundation
        required:
            - name
            - domain
...
    QuerySvcQueryResourcesCountResponseBody:
        title: QuerySvcQueryResourcesCountResponseBody
        type: object
        properties:
            count:
                type: integer
                description: Count of resources found
                example: 1234
                format: int64
            has_more:
                type: boolean
                description: 'True if count is not guaranteed to be exhaustive: client should request a narrower query'
                example: false
        example:
            count: 1234
            has_more: false
        required:
            - count
            - has_more
...
    QuerySvcSuggestOrgsResponseBody:
        title: QuerySvcSuggestOrgsResponseBody
        type: object
        properties:
            suggestions:
                type: array
                items:
                    $ref: '#/definitions/OrganizationSuggestion'
                description: Organization suggestions
                example:
                    - domain: linuxfoundation.org
                      logo: https://example.com/logo.png
                      name: Linux Foundation
                    - domain: linuxfoundation.org
                      logo: https://example.com/logo.png
                      name: Linux Foundation
        example:
            suggestions:
                - domain: linuxfoundation.org
                  logo: https://example.com/logo.png
                  name: Linux Foundation
                - domain: linuxfoundation.org
                  logo: https://example.com/logo.png
                  name: Linux Foundation
                - domain: linuxfoundation.org
                  logo: https://example.com/logo.png
                  name: Linux Foundation
                - domain: linuxfoundation.org
                  logo: https://example.com/logo.png
                  name: Linux Foundation
        required:
            - suggestions
...
    ServiceUnavailableError:
        title: ServiceUnavailableError
        type: object
        properties:
            message:
                type: string
                description: Error message
                example: The service is unavailable.
        description: Service unavailable
        example:
            message: The service is unavailable.
        required:
            - message
securityDefinitions:
    jwt_header_Authorization:
        type: apiKey
        description: Heimdall authorization
        name: Authorization
        in: header